WIP: Created new tests to validate actions of a robot
