+++ src/grid-mapper.js
@@ -1,32 +1,10 @@
-function mapGirdCoordinates(firstLineInput) {
-    const coordinate = firstLineInput.trim()
-        .split(' ')
-        .map(number => parseInt(number, 10));
-
-    return {
-        x: coordinate[0],
-        y: coordinate[1]
-    };
-}
-
-function mapGridPosition(secondLineInput) {
-    const position = secondLineInput.trim().split(' ');
-
-    return {
-        x: parseInt(position[0], 10),
-        y: parseInt(position[1], 10),
-        orientation: position[2]
-    };
-}
-
-function mapInstructions(thirdLineInput) {
-    return thirdLineInput.trim().split('');
-}
+const gridArea = require('../src/grid')
+const robot = require('../src/robot')
 
 exports.mapGrid = function (input) {
     const grid = input.split('\n');
 
-    const gridMap = mapGirdCoordinates(grid[0]);
+    const gridMap = gridArea.mapGirdCoordinates(grid[0]);
     const area = {
         top: gridMap.y,
         right: gridMap.x,
@@ -37,8 +15,8 @@ exports.mapGrid = function (input) {
     const robots = [];
     for (let i = 1; i < grid.length; i += 2) {
         robots.push({
-            position: mapGridPosition(grid[i]),
-            instructions: mapInstructions(grid[i + 1])
+            position: robot.mapGridPosition(grid[i]),
+            instructions: robot.mapInstructions(grid[i + 1])
         });
     }
